+++ R/shoji-catalog.R
@@ -8,15 +8,6 @@ setMethod("initialize", "ShojiCatalog", init.ShojiCatalog)
 is.shojiCatalog <- function (x) inherits(x, "ShojiCatalog")
 
 setIndexSlot <- function (x, i, value) {
-    index(x) <- lapply(x, function (a) {
-        a[[i]] <- value
-        return(a)
-    })
-    crPATCH(self(x), body=toJSON(index(x)))
-    return(x)
-}
-
-mapSetIndexSlot <- function (x, i, value) {
     if (length(value) == 1) value <- rep(value, length(x))
     stopifnot(length(x) == length(value))
     
@@ -30,6 +21,7 @@ mapSetIndexSlot <- function (x, i, value) {
         ## Make sure certain fields are [] in the JSON
         ensure <- c("subvariables")
         payload <- lapply(index(x)[to.update], function (p) {
+            p <- p[i] ## Let's only PATCH the field we're changing
             these <- intersect(ensure, names(p))
             if (length(these)) p[these] <- lapply(p[these], I)
             return(p)