+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.2.3
+* Fix bug in updating an array with only one subvariable.
 
 ### crunch 1.2.2
 * Add `types` method to VariableCatalog.